+++ docs/docs/changelog.md
@@ -7,6 +7,8 @@
 - Fixed issues with releases performence in `calendar` widget 
 - Made image fetching in `currently playing` more reliable
 - Added a function in `calendar` widget when showing releases to also show release type
+- Added a function in `calendar` widget to also show current state of the media e.g. grabbed
+- Fixed issue with `search` widget highlighting
 
 # Changes for 2.4.0
 - Added Brave Search as an autocompletion engine and normal one